requirements.txt

torch==2.6.0
torchvision==0.21.0
numpy==2.1.2
pandas==2.3.2
matplotlib==3.10.0
scikit-learn==1.6.1
tqdm==4.67.1
opencv-python==4.11.0.86
Pillow==11.1.0
einops==0.8.1
